requirements.txt
Django==4.1.4
django-browser-reload==1.6.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0